Use characters list in participants modal
